versão beta

+++ EmployeeSharp.Infra/Data/Repositories/ColaboradorRepository.cs
@@ -1,11 +1,6 @@
 ﻿using EmployeeSharp.Domain.Entities;
 using EmployeeSharp.Domain.Interfaces;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EmployeeSharp.Infra.Data.Repositories
 {
@@ -18,39 +13,46 @@ namespace EmployeeSharp.Infra.Data.Repositories
             _context = context;
         }
 
-        public async Task<Colaborador> GetByIdAsync(int id)
-        {
-            return await _context.Colaboradores.Include(c => c.Cargo).FirstOrDefaultAsync(c => c.Id == id);
-        }
-
         public async Task<IEnumerable<Colaborador>> GetAllAsync()
         {
-            return await _context.Colaboradores.Include(c => c.Cargo).ToListAsync();
+            return await _context.Colaboradores.ToListAsync();
+        }
+
+        public async Task<Colaborador> GetByIdAsync(int id)
+        {
+            return await _context.Colaboradores.FindAsync(id);
         }
 
         public async Task<IEnumerable<Colaborador>> SearchByNameAsync(string nome)
         {
-            return await _context.Colaboradores.Include(c => c.Cargo)
-                .Where(c => c.Nome.Contains(nome)).ToListAsync();
+            return await _context.Colaboradores.Where(c => c.Nome.Contains(nome)).ToListAsync();
         }
 
         public async Task AddAsync(Colaborador colaborador)
         {
-            await _context.Colaboradores.AddAsync(colaborador);
+            _context.Colaboradores.Add(colaborador);
             await _context.SaveChangesAsync();
         }
 
         public async Task UpdateAsync(Colaborador colaborador)
         {
-            _context.Colaboradores.Update(colaborador);
+            _context.Entry(colaborador).State = EntityState.Modified;
             await _context.SaveChangesAsync();
         }
 
         public async Task DeleteAsync(int id)
         {
-            var colaborador = await GetByIdAsync(id);
-            _context.Colaboradores.Remove(colaborador);
-            await _context.SaveChangesAsync();
+            var colaborador = await _context.Colaboradores.FindAsync(id);
+            if (colaborador != null)
+            {
+                _context.Colaboradores.Remove(colaborador);
+                await _context.SaveChangesAsync();
+            }
+        }
+
+        public async Task<IEnumerable<Colaborador>> GetByCargoIdAsync(int cargoId)
+        {
+            return await _context.Colaboradores.Where(c => c.CargoId == cargoId).ToListAsync();
         }
     }
 }
